Utils: Refactor into single utils module
Move checks, decorators and paths to utils module.
